+++ common/run.sh
@@ -11,7 +11,18 @@ add2iptables "INPUT -p udp --dport 3478 -j ACCEPT"
 add2iptables "INPUT -p udp --dport 53 -j ACCEPT"
 add2iptables "INPUT -p tcp --dport 80 -j ACCEPT"
 add2iptables "INPUT -p tcp --dport 22 -j ACCEPT"
+
+for PORT in ${HTTP_PORTS//,/ };	do 
+  add2iptables "INPUT -p tcp --dport $PORT -j ACCEPT"
+done
+for PORT in ${TLS_PORTS//,/ };	do 
+  add2iptables "INPUT -p tcp --dport $PORT -j ACCEPT"
+done
+
 mkdir -p /etc/iptables/
+
+
+
  if [[ $ENABLE_FIREWALL == true ]]; then
   iptables -P INPUT DROP
   iptables-save > /etc/iptables/rules.v4