+++ hiddify-panel/backup.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+cd $( dirname -- "$0"; )
+function main(){
+    python3 -m hiddifypanel backup
+}
+main |& tee -a ../log/system/backup.log