+++ hiddify-panel/install.sh
@@ -41,3 +41,6 @@ service cron reload
 systemctl start hiddify-panel.service
 systemctl status hiddify-panel.service
 
+
+echo "0 */6 * * * root $(pwd)/backup.sh" > /etc/cron.d/hiddify_auto_backup
+service cron reload